lpcollection/requirements.txt
router
djongo
discogs_client
django-cors-headers
dnspython
django_rest_framework